requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==1.10.15
python-dotenv==1.0.1
supabase==2.5.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
midtransclient==1.4.1
httpx==0.27.0
python-multipart